Update patch set 1
Patch Set 1: Code-Review-2

(12 comments)

It is good for a proof of concept, but we need to structure it more carefully before integrating. I'll propose a structure that allows testing, so we can add it too. In the meantime there are already a few points to consider and change in the comments.

Patch-set: 1
Reviewer: Thiago Figueredo Cardoso <1013086@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-2